# api/swagger.yaml
basePath: /
definitions:
...
  model.AnimeSearch:
    properties:
      endDate:
        $ref: '#/definitions/model.Date'
      episode:
        type: integer
      id:
        type: integer
      image:
        type: string
      member:
        type: integer
      rated:
        type: string
      score:
        type: number
      startDate:
        $ref: '#/definitions/model.Date'
      summary:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  model.CharacterSearch:
    properties:
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      nickname:
        type: string
    type: object
...
  model.ClubSearch:
    properties:
      creator:
        type: string
      id:
        type: integer
      image:
        type: string
      member:
        type: integer
      name:
        type: string
      summary:
        type: string
    type: object
  model.Date:
    properties:
      day:
        type: integer
      month:
        type: integer
      year:
        type: integer
    type: object
...
  model.MangaSearch:
    properties:
      chapter:
        type: integer
      endDate:
        $ref: '#/definitions/model.Date'
      id:
        type: integer
      image:
        type: string
      member:
        type: integer
      score:
        type: number
      startDate:
        $ref: '#/definitions/model.Date'
      summary:
        type: string
      title:
        type: string
      type:
        type: string
      volume:
        type: integer
    type: object
...
  model.PeopleSearch:
    properties:
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      nickname:
        type: string
    type: object
...
  model.UserSearch:
    properties:
      image:
        type: string
      lastOnline:
        type: string
      username:
        type: string
    type: object
...
  utils.Response:
    properties:
      data:
        type: object
      message:
        type: string
      meta:
        type: object
      status:
        type: integer
    type: object
info:
  contact:
    email: axel.rl.404@gmail.com
    name: Axel
    url: https://github.com/rl404
  description: Scraping/parsing MyAnimeList website to a useful and easy-to-use data.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Malscraper API
paths:
...
  /search/anime:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime title
        in: query
        name: query
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Anime type (1=TV, 2=OVA, 3=Movie, 4=Special, 5=ONA, 6=Music)
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        in: query
        name: type
        type: integer
      - description: Score
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
        - 9
        - 10
        in: query
        name: score
        type: integer
      - description: Anime airing status (1=on-going, 2=finished, 3=upcoming)
        enum:
        - 0
        - 1
        - 2
        - 3
        in: query
        name: status
        type: integer
      - description: Producer ID
        in: query
        name: producer
        type: integer
      - description: Anime rating (1=G, 2=PG, 3=PG13, 4=R17, 5=R, 6=RX)
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        in: query
        name: rating
        type: integer
      - description: Anime airing start date (yyyy-mm-dd)
        in: query
        name: start
        type: string
      - description: Anime airing end date (yyyy-mm-dd)
        in: query
        name: end
        type: string
      - description: Exlude genre?
        in: query
        name: xgenre
        type: boolean
      - description: Genre ID
        in: query
        items:
          type: integer
        name: genre
        type: array
      - description: Anime first letter
        in: query
        name: letter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.AnimeSearch'
                  type: array
              type: object
      summary: Search anime
      tags:
      - search
  /search/character:
    get:
      consumes:
      - application/json
      parameters:
      - description: Character name
        in: query
        name: query
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.CharacterSearch'
                  type: array
              type: object
      summary: Search character
      tags:
      - search
  /search/club:
    get:
      consumes:
      - application/json
      parameters:
      - description: Club name
        in: query
        name: query
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Category (1=anime, 2=convention, 3=actor & artist, 4=character, 5=company, 6=game, 7=japan, 8=city, 9=music, 10=manga, 11=school, 12=other)
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
        - 9
        - 10
        - 11
        - 12
        in: query
        name: category
        type: integer
      - description: Sort (1=name, 2=comment, 3=post, 5=member)
        enum:
        - 0
        - 1
        - 2
        - 3
        - 5
        in: query
        name: sort
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ClubSearch'
                  type: array
              type: object
      summary: Search club
      tags:
      - search
  /search/manga:
    get:
      consumes:
      - application/json
      parameters:
      - description: Manga title
        in: query
        name: query
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Manga type (1=Manga, 2=Light Novel, 3=One-shot, 4=Doujinshi, 5=Manhwa, 6=Manhua, 8=Novel)
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 8
        in: query
        name: type
        type: integer
      - description: Score
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
        - 9
        - 10
        in: query
        name: score
        type: integer
      - description: Manga publishing status (1=on-going, 2=finished, 3=upcoming, 4=hiatus, 5=discontinued)
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        in: query
        name: status
        type: integer
      - description: Magazine ID
        in: query
        name: magazine
        type: integer
      - description: Manga publishing start date (yyyy-mm-dd)
        in: query
        name: start
        type: string
      - description: Manga publishing end date (yyyy-mm-dd)
        in: query
        name: end
        type: string
      - description: Exlude genre?
        in: query
        name: xgenre
        type: boolean
      - description: Genre ID
        in: query
        items:
          type: integer
        name: genre
        type: array
      - description: Manga first letter
        in: query
        name: letter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.MangaSearch'
                  type: array
              type: object
      summary: Search manga
      tags:
      - search
  /search/people:
    get:
      consumes:
      - application/json
      parameters:
      - description: People name
        in: query
        name: query
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.PeopleSearch'
                  type: array
              type: object
      summary: Search people
      tags:
      - search
  /search/user:
    get:
      consumes:
      - application/json
      parameters:
      - description: User name
        in: query
        name: query
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Location
        in: query
        name: location
        type: string
      - description: Minimum age
        in: query
        name: minAge
        type: integer
      - description: Maximum age
        in: query
        name: maxAge
        type: integer
      - description: Gender (1=male, 2=female, 3=non-binary)
        enum:
        - "0"
        - "1"
        - "2"
        - "3"
        in: query
        name: gender
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserSearch'
                  type: array
              type: object
      summary: Search user
      tags:
      - search
...
schemes:
- http
- https
swagger: "2.0"